+++ PHP/add_update/remove_person.php
@@ -2,7 +2,7 @@
     include "../config_DB.php";
     $id=$_SESSION['session'];
     $id_person=$_POST['user_id'];
-    if($id != $id_person){
+    if($id != $id_person && isset($_SESSION['session'])){
         $sql = "Delete from prirad_funkcia where os_udaje_rod_cislo='" . $id_person . "'";
         $con->query($sql);
         $sql = "Delete from login where os_udaje_rod_cislo='" . $id_person . "'";
